Embedded HTML documents
IE 11 was having issues with the (wrong) way in which the Ember app was being embedded - there were multiple HTML and body tags, which probably caused some issue with selectors, it’s not clear. This restructures the embedding so that the Ember index.html does most of it, the rails layout is simply blank yielding, and the rails template itself renders the Rails partials accordingly (yes, there is a mixture of Rails partials and Ember application both doing javascript things, but that’s a separate conversation). This means no more warnings in IE 11 about duplicate HTML and Body tags.